+++ src/interval.rs
@@ -1,35 +1,35 @@
-use core::f32;
+use core::f64;
 
 pub struct Interval {
-    pub min: f32,
-    pub max: f32,
+    pub min: f64,
+    pub max: f64,
 }
 
 impl Interval {
     pub fn new() -> Interval {
         Interval {
-            min: f32::NEG_INFINITY,
-            max: f32::INFINITY,
+            min: f64::NEG_INFINITY,
+            max: f64::INFINITY,
         }
     }
 
-    pub fn with_values(min: f32, max: f32) -> Interval {
+    pub fn with_values(min: f64, max: f64) -> Interval {
         Interval { min, max }
     }
 
-    pub fn size(&self) -> f32 {
+    pub fn size(&self) -> f64 {
         self.max - self.min
     }
 
-    pub fn contians(&self, value: f32) -> bool {
+    pub fn contians(&self, value: f64) -> bool {
         self.min <= value && value <= self.max
     }
 
-    pub fn surrounds(&self, value: f32) -> bool {
+    pub fn surrounds(&self, value: f64) -> bool {
         self.min < value && value < self.max
     }
 
-    pub fn clamp(&self, value: f32) -> f32 {
+    pub fn clamp(&self, value: f64) -> f64 {
         if value < self.min {
             return self.min;
         } else if value > self.max {
@@ -40,12 +40,12 @@ impl Interval {
     }
 
     pub const EMPTY: Interval = Interval {
-        min: f32::INFINITY,
-        max: f32::NEG_INFINITY,
+        min: f64::INFINITY,
+        max: f64::NEG_INFINITY,
     };
 
     pub const UNIVERSE: Interval = Interval {
-        min: f32::NEG_INFINITY,
-        max: f32::INFINITY,
+        min: f64::NEG_INFINITY,
+        max: f64::INFINITY,
     };
 }
