+++ src/ray.rs
@@ -28,7 +28,7 @@ impl Ray {
         self.dir
     }
 
-    pub fn at(&self, t: f32) -> Point3 {
+    pub fn at(&self, t: f64) -> Point3 {
         return self.origin + t * self.dir;
     }
 }
